requirements.txt
streamlit
whisper
sounddevice
scipy
numpy
openai
python-dotenv
requests
google-generativeai